Refactor node crypto to Web Crypto API
Node crypto is not supported on edge runtimes in Next.js and Web
Crypto API is async. This bring changes in the library, such as
converting the library into a Asynchronous including all functions
such as encode_jwt, verify_jwt etc.

Jest testEnvironment is changed to `jsdom` from `node` since, the
runtime is changed from node to browser.

Signed-off-by: Pradyumna Krishna <git@onpy.in>